ree_fs: allow seeking past of the end of file
OP-TEE (or TAs) sometimes try to seek past of the file. This is normal
and is allowed by POSIX. Problem is that Zephyr FS API does not allow
this.

So wee need to mimic POSIX behavior at our side. In case of "read"
operation, we will just report that zero bytes were read, the same as
POSIX `pread` function does. For "write" operation it is a bit harder,
because we need to pad the file with zeroes up to the desired position.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>